python_agents/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
google-adk>=0.5.0
google-cloud-bigquery>=3.0.0
python-dotenv>=1.0.0 